feat: calcule l'égilibilité des parcours moyennes et grandes entreprises privées en France
